test: fix eth_addr_reg test case
