+++ polyjuice-tests/src/test_cases/eth_addr_reg.rs
@@ -1,10 +1,11 @@
 use crate::helper::{
     build_eth_l2_script, new_block_info, setup, update_eth_address_registry, CKB_SUDT_ACCOUNT_ID,
-    ETH_ADDRESS_REGISTRY_PROGRAM_CODE_HASH, L2TX_MAX_CYCLES,
+    ETH_ADDRESS_REGISTRY_ACCOUNT_ID, L2TX_MAX_CYCLES,
 };
 use gw_common::state::State;
 use gw_generator::traits::StateExt;
 use gw_store::chain_view::ChainView;
+use gw_store::traits::chain_store::ChainStore;
 use gw_types::{
     core::ScriptHashType,
     packed::{RawL2Transaction, Script},
@@ -56,54 +57,11 @@ impl GwToEthArgsBuilder {
     }
 }
 
-#[derive(Default)]
-struct SetMappingArgsBuilder {
-    method: u32,
-    gw_script_hash: [u8; 32],
-    fee_struct: gw_types::packed::Fee,
-}
-impl SetMappingArgsBuilder {
-    /// Set the SetMappingArgs builder's method.
-    fn method(mut self, method: u32) -> Self {
-        self.method = method;
-        self
-    }
-    /// Set the SetMappingArgs builder's gw script hash.
-    fn gw_script_hash(mut self, gw_script_hash: [u8; 32]) -> Self {
-        self.gw_script_hash = gw_script_hash;
-        self
-    }
-    /// Set the SetMappingArgs builder's fee struct.
-    fn fee_struct(mut self, fee_sudt_id: u32, fee_amount: u128) -> Self {
-        self.fee_struct = gw_types::packed::Fee::new_builder()
-            .amount(fee_amount.pack())
-            .sudt_id(fee_sudt_id.pack())
-            .build();
-        self
-    }
-    fn build(self) -> Vec<u8> {
-        let mut output: Vec<u8> = vec![0u8; 4];
-        output[0..4].copy_from_slice(&self.method.to_le_bytes()[..]);
-        output.extend(self.gw_script_hash);
-        output.extend(self.fee_struct.as_slice());
-        output
-    }
-}
-
 #[test]
 fn test_eth_to_gw() {
     let (store, mut state, generator, _creator_account_id) = setup();
 
     // init accounts
-    let eth_addr_reg_account_id = state
-        .create_account_from_script(
-            Script::new_builder()
-                .code_hash(ETH_ADDRESS_REGISTRY_PROGRAM_CODE_HASH.clone().pack())
-                .args([0u8; 32].to_vec().pack())
-                .hash_type(ScriptHashType::Type.into())
-                .build(),
-        )
-        .expect("create eth_addr_reg_account");
     let a_id = state
         .create_account_from_script(
             Script::new_builder()
@@ -130,12 +88,12 @@ fn test_eth_to_gw() {
         .build();
     let raw_tx = RawL2Transaction::new_builder()
         .from_id(a_id.pack())
-        .to_id(eth_addr_reg_account_id.pack())
+        .to_id(ETH_ADDRESS_REGISTRY_ACCOUNT_ID.pack())
         .args(args.pack())
         .build();
     let block_info = new_block_info(a_id, 1, 0);
-    let tip_block_hash = store.get_tip_block_hash().unwrap();
     let db = store.begin_transaction();
+    let tip_block_hash = db.get_tip_block_hash().unwrap();
     let run_result = generator
         .execute_transaction(
             &ChainView::new(&db, tip_block_hash),
@@ -160,15 +118,6 @@ fn test_gw_to_eth() {
     let (store, mut state, generator, _creator_account_id) = setup();
 
     // init accounts
-    let eth_addr_reg_account_id = state
-        .create_account_from_script(
-            Script::new_builder()
-                .code_hash(ETH_ADDRESS_REGISTRY_PROGRAM_CODE_HASH.clone().pack())
-                .args([0u8; 32].to_vec().pack())
-                .hash_type(ScriptHashType::Type.into())
-                .build(),
-        )
-        .expect("create eth_addr_reg_account");
     let a_id = state
         .create_account_from_script(
             Script::new_builder()
@@ -190,7 +139,7 @@ fn test_gw_to_eth() {
         .build();
     let raw_l2tx = RawL2Transaction::new_builder()
         .from_id(a_id.pack())
-        .to_id(eth_addr_reg_account_id.pack())
+        .to_id(ETH_ADDRESS_REGISTRY_ACCOUNT_ID.pack())
         .args(args.pack())
         .build();
     let db = store.begin_transaction();
@@ -214,17 +163,6 @@ fn test_set_mapping_by_contract() {
     let (store, mut state, generator, _creator_account_id) = setup();
 
     // init accounts
-    let eth_addr_reg_account_id = state
-        .create_account_from_script(
-            Script::new_builder()
-                .code_hash(ETH_ADDRESS_REGISTRY_PROGRAM_CODE_HASH.clone().pack())
-                .args([0u8; 32].to_vec().pack())
-                .hash_type(ScriptHashType::Type.into())
-                .build(),
-        )
-        .expect("create eth_addr_reg_account");
-    println!("eth_addr_reg_account_id {}", eth_addr_reg_account_id);
-
     let from_id = state
         .create_account_from_script(
             Script::new_builder()
@@ -262,30 +200,14 @@ fn test_set_mapping_by_contract() {
     );
 
     // update_eth_address_registry
-    let args = SetMappingArgsBuilder::default()
-        .method(2u32)
-        .gw_script_hash(eth_eoa_account_script_hash)
-        .fee_struct(CKB_SUDT_ACCOUNT_ID, 1000)
-        .build();
-    let raw_l2tx = RawL2Transaction::new_builder()
-        .from_id(eth_eoa_account_id.pack())
-        .to_id(eth_addr_reg_account_id.pack())
-        .args(args.pack())
-        .build();
-    let db = store.begin_transaction();
-    let tip_block_hash = store.get_tip_block_hash().unwrap();
-    let block_info = new_block_info(from_id, 1, 0);
-    let run_result = generator
-        .execute_transaction(
-            &ChainView::new(&db, tip_block_hash),
-            &state,
-            &block_info,
-            &raw_l2tx,
-            L2TX_MAX_CYCLES,
-        )
-        .expect("execute Godwoken contract");
-    state.apply_run_result(&run_result).expect("update state");
-    assert_eq!(run_result.exit_code, 0);
+    crate::helper::eth_address_regiser(
+        &store,
+        &mut state,
+        &generator,
+        1,
+        eth_eoa_account_id,
+        eth_eoa_account_script_hash.into(),
+    );
 
     // check result: eth_address -> gw_script_hash
     let args = EthToGwArgsBuilder::default()
@@ -294,7 +216,7 @@ fn test_set_mapping_by_contract() {
         .build();
     let raw_l2tx = RawL2Transaction::new_builder()
         .from_id(from_id.pack())
-        .to_id(eth_addr_reg_account_id.pack())
+        .to_id(ETH_ADDRESS_REGISTRY_ACCOUNT_ID.pack())
         .args(args.pack())
         .build();
     let block_info = new_block_info(from_id, 2, 0);
@@ -318,11 +240,11 @@ fn test_set_mapping_by_contract() {
         .build();
     let raw_l2tx = RawL2Transaction::new_builder()
         .from_id(from_id.pack())
-        .to_id(eth_addr_reg_account_id.pack())
+        .to_id(ETH_ADDRESS_REGISTRY_ACCOUNT_ID.pack())
         .args(args.pack())
         .build();
     let db = store.begin_transaction();
-    let tip_block_hash = store.get_tip_block_hash().unwrap();
+    let tip_block_hash = db.get_tip_block_hash().unwrap();
     let block_info = new_block_info(from_id, 3, 0);
     let run_result = generator
         .execute_transaction(
@@ -336,4 +258,8 @@ fn test_set_mapping_by_contract() {
     assert_eq!(run_result.return_data, eth_address);
 
     // TODO: test conatract account
+    // [contract_account_script_hash <-> short_account_script_hash as eth_address]
+    // {
+    //     let creator_script_hash = state.get_script_hash(_creator_account_id);
+    // }
 }
